app_api/requirements.txt
avinit[png]
a2wsgi
boto3==1.26.82
coverage==7.2.1
gunicorn
httpx
firebase-admin==6.1.0
kink
motor==3.1.1
pytest==7.2.1
pytest-asyncio
PyJWT==2.6.0
python-bsonjs
pyyaml
Pillow==9.4.0
pydantic==1.10.5
pymongo==4.3.3
python-multipart==0.0.6
rq==1.13.0
requests==2.28.2
redis==4.5.1
starlette[full]==0.25.0
sentry-sdk[starlette]==1.16.0
uvicorn==0.20.0
